Revert "Merge branch 'update-arm-simd-intrinsic' of https://github.com/sergey-grovety/tvm into issue8717-x86-DwsConv2d-schedule"
This reverts commit e927567058403bcc9e4fdc3d24828b3dcd6a661b, reversing
changes made to 0ccb5a01495d02f521eea2af9efa6a3153c4f72b.
