+++ apps/microtvm/reference-vm/zephyr/base-box/base_box_test.sh
@@ -37,5 +37,3 @@ if [ $board == "stm32f746xx" ]; then
 else
     pytest tests/micro/zephyr/test_zephyr_aot.py --zephyr-board=${board}
 fi
-
-pytest tests/micro/zephyr/test_zephyr_armv7m.py --zephyr-board=${board}
