+++ python/tvm/topi/arm_cpu/cortex_m7/micro_kernel/gemm.py
@@ -46,8 +46,9 @@ def intrin_gemm_MxKxN(M, K, N, in_dtype, out_dtype):
         K = K.value
     if isinstance(N, tvm.tir.IntImm):
         N = N.value
+    assert K % 4 == 0
     # TODO(weberlo, areusch): support more dtypes?
-    assert in_dtype in ("int8", "int16")
+    assert in_dtype == "int8"
     assert out_dtype == "int32"
     A = te.placeholder((M, K), name="a", dtype=in_dtype)
     B = te.placeholder((N, K), name="b", dtype=in_dtype)
@@ -70,14 +71,13 @@ def intrin_gemm_MxKxN(M, K, N, in_dtype, out_dtype):
     def intrin_func(ins, outs):
         aa, bb = ins
         cc = outs[0]
-        gemm_func_prefix = "gemm" if in_dtype == "int8" else "gemm16"
 
         def _reduce_update():
             ib = tvm.tir.ir_builder.create()
             ib.emit(
                 tvm.tir.call_extern(
                     "int32",
-                    f"{gemm_func_prefix}_{M}x{K}x{N}_update_{uniq_id}",
+                    f"gemm_{M}x{K}x{N}_update_{uniq_id}",
                     aa.access_ptr("r"),
                     bb.access_ptr("r"),
                     cc.access_ptr("w"),
@@ -102,7 +102,7 @@ def intrin_gemm_MxKxN(M, K, N, in_dtype, out_dtype):
             ib.emit(
                 tvm.tir.call_extern(
                     "int32",
-                    f"{gemm_func_prefix}_{M}x{K}x{N}_body_{uniq_id}",
+                    f"gemm_{M}x{K}x{N}_body_{uniq_id}",
                     aa.access_ptr("r"),
                     bb.access_ptr("r"),
                     cc.access_ptr("w"),
@@ -122,7 +122,7 @@ def intrin_gemm_MxKxN(M, K, N, in_dtype, out_dtype):
 def gemm_MxKxN_impl(M, K, N, uniq_id):
     """Emit C code for gemm impl."""
     # TODO(weberlo, areusch): are there any SIMD tricks to zero out arrays quickly?
-    # aa_pad_size = M * K
+    aa_pad_size = M * K
     bb_pad_size = N * K
     # code reference: CMSIS-NN paper (https://arxiv.org/abs/1801.06601)
     cc_code = f"""
@@ -132,97 +132,32 @@ extern "C"
 #include <arm_math.h>
 #include <arm_nnsupportfunctions.h>
 
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm_{M}x{N}_body_rest_{uniq_id}(
-    int K,
-    int8_t *aa, int8_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{
-  int k_base = (K / 4) * 4;
-  switch ( K % 4 ) {{
-  case 1:
-    for (int i = 0; i < {M}; i++) {{
-      for (int j = 0; j < {N}; j++) {{
-        int8_t *a_ptr = &aa[i * A_stride + k_base];
-        int8_t *b_ptr = &bb[j * B_stride + k_base];
-        cc[i * C_stride + j] = (int32_t) a_ptr[0] * (int32_t) b_ptr[0];
-      }}
-    }}
-    break;
-  case 2:
-    for (int i = 0; i < {M}; i++) {{
-      for (int j = 0; j < {N}; j++) {{
-        int8_t *a_ptr = &aa[i * A_stride + k_base];
-        int8_t *b_ptr = &bb[j * B_stride + k_base];
-        cc[i * C_stride + j] =   (int32_t) a_ptr[0] * (int32_t) b_ptr[0]
-                               + (int32_t) a_ptr[1] * (int32_t) b_ptr[1];
-      }}
-    }}
-    break;
-  case 3:
-    for (int i = 0; i < {M}; i++) {{
-      for (int j = 0; j < {N}; j++) {{
-        int8_t *a_ptr = &aa[i * A_stride + k_base];
-        int8_t *b_ptr = &bb[j * B_stride + k_base];
-        cc[i * C_stride + j] =   (int32_t) a_ptr[0] * (int32_t) b_ptr[0]
-                               + (int32_t) a_ptr[1] * (int32_t) b_ptr[1]
-                               + (int32_t) a_ptr[2] * (int32_t) b_ptr[2];
-      }}
-    }}
-    break;
-  }}
-  return 0;
-}}
-
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm_{M}x{K}x{N}_body_loop_{uniq_id}(
-    int8_t *aa, int8_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{
-  for (int i = 0; i < {M}; i++) {{
-    for (int j = 0; j < {N}; j++) {{
-      int32_t sum = 0;
-      for (int l = 0; l < {K}; l++) {{
-        sum += (int32_t) aa[i*A_stride + l] * (int32_t) bb[j*B_stride + l];
-      }}
-      // NOTE: this is the line where `*_body` differs from `*_update`. here
-      // we're *setting* the result, instead of accumulating, because we know
-      // the `i` and `j` itervars span their entire respective axes.
-      cc[i*C_stride + j] = sum;
-    }}
-  }}
-  return 0;
-}}
-
-#ifdef __cplusplus
-extern "C"
-#endif
 __STATIC_FORCEINLINE int32_t gemm_{M}x{K}x{N}_body_{uniq_id}(
     int8_t *aa, int8_t *bb, int32_t *cc,
     int A_stride, int B_stride, int C_stride) {{
+  int16_t aa_pad[{aa_pad_size}];
   int16_t bb_pad[{bb_pad_size}];
 
-  if ( {M} < 16 || {N} < 16 )
-    return gemm_{M}x{K}x{N}_body_loop_{uniq_id}(aa, bb, cc, A_stride, B_stride, C_stride);
+  for (int i = 0; i < {M}; i++) {{
+    for (int j = 0; j < {K} / 4; j++) {{
+      read_and_pad(&aa[i*A_stride + j*4], (int32_t*) &aa_pad[i*{K} + j*4], (int32_t*) &aa_pad[i*{K} + j*4 + 2]);
+    }}
+  }}
 
-  for (int i = 0; i < {N}; i++)
-    for (int j = 0; j < {K} / 4; j++)
+  for (int i = 0; i < {N}; i++) {{
+    for (int j = 0; j < {K} / 4; j++) {{
       read_and_pad(&bb[i*B_stride + j*4], (int32_t*) &bb_pad[i*{K} + j*4], (int32_t*) &bb_pad[i*{K} + j*4 + 2]);
+    }}
+  }}
 
   for (int i = 0; i < {M}; i++) {{
-    int16_t aa_pad_line[{K}];
-    for (int l = 0; l < {K} / 4; l++)
-      read_and_pad(&aa[i*A_stride + l*4], (int32_t*) &aa_pad_line[l*4], (int32_t*) &aa_pad_line[l*4 + 2]);
-
     for (int j = 0; j < {N}; j++) {{
-      int32_t *aa_ptr = (int32_t *) aa_pad_line;
-      int32_t *bb_ptr = (int32_t *) &bb_pad[j*{K}];
       int32_t sum = 0;
-      for (int l = 0; l < 2 * ({K} / 4); l++) {{
-        sum = __SMLAD(*aa_ptr, *bb_ptr, sum);
-        ++ aa_ptr; ++ bb_ptr;
+      for (int l = 0; l < {K} / 2; l++) {{
+        sum = __SMLAD(
+          *((int32_t*) &aa_pad[i*{K} + l*2]),
+          *((int32_t*) &bb_pad[j*{K} + l*2]),
+          sum);
       }}
       // NOTE: this is the line where `*_body` differs from `*_update`. here
       // we're *setting* the result, instead of accumulating, because we know
@@ -231,71 +166,6 @@ __STATIC_FORCEINLINE int32_t gemm_{M}x{K}x{N}_body_{uniq_id}(
     }}
   }}
 
-  if ( {K} % 4 != 0 )
-    gemm_{M}x{N}_body_rest_{uniq_id}({K}, aa, bb, cc, A_stride, B_stride, C_stride);
-
-  return 0;
-}}
-
-
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm_{M}x{N}_update_rest_{uniq_id}(
-    int K,
-    int8_t *aa, int8_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{
-  int k_base = (K / 4) * 4;
-  switch ( K % 4 ) {{
-  case 1:
-    for (int i = 0; i < {M}; i++) {{
-      for (int j = 0; j < {N}; j++) {{
-        int8_t *a_ptr = &aa[i * A_stride + k_base];
-        int8_t *b_ptr = &bb[j * B_stride + k_base];
-        cc[i * C_stride + j] += (int32_t) a_ptr[0] * (int32_t) b_ptr[0];
-      }}
-    }}
-    break;
-  case 2:
-    for (int i = 0; i < {M}; i++) {{
-      for (int j = 0; j < {N}; j++) {{
-        int8_t *a_ptr = &aa[i * A_stride + k_base];
-        int8_t *b_ptr = &bb[j * B_stride + k_base];
-        cc[i * C_stride + j] +=   (int32_t) a_ptr[0] * (int32_t) b_ptr[0]
-                                + (int32_t) a_ptr[1] * (int32_t) b_ptr[1];
-      }}
-    }}
-    break;
-  case 3:
-    for (int i = 0; i < {M}; i++) {{
-      for (int j = 0; j < {N}; j++) {{
-        int8_t *a_ptr = &aa[i * A_stride + k_base];
-        int8_t *b_ptr = &bb[j * B_stride + k_base];
-        cc[i * C_stride + j] +=   (int32_t) a_ptr[0] * (int32_t) b_ptr[0]
-                                + (int32_t) a_ptr[1] * (int32_t) b_ptr[1]
-                                + (int32_t) a_ptr[2] * (int32_t) b_ptr[2];
-      }}
-    }}
-    break;
-  }}
-  return 0;
-}}
-
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm_{M}x{K}x{N}_update_loop_{uniq_id}(
-    int8_t *aa, int8_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{
-  for (int i = 0; i < {M}; i++) {{
-    for (int j = 0; j < {N}; j++) {{
-      int32_t sum = 0;
-      for (int l = 0; l < {K}; l++) {{
-        sum += (int32_t) aa[i*A_stride + l] * (int32_t) bb[j*B_stride + l];
-      }}
-      cc[i*C_stride + j] += sum;
-    }}
-  }}
   return 0;
 }}
 
@@ -305,178 +175,36 @@ extern "C"
 __STATIC_FORCEINLINE int32_t gemm_{M}x{K}x{N}_update_{uniq_id}(
     int8_t *aa, int8_t *bb, int32_t *cc,
     int A_stride, int B_stride, int C_stride) {{
+  int16_t aa_pad[{aa_pad_size}];
   int16_t bb_pad[{bb_pad_size}];
 
-  if ( {M} < 16 || {N} < 16 )
-    return gemm_{M}x{K}x{N}_update_loop_{uniq_id}(aa, bb, cc, A_stride, B_stride, C_stride);
+  for (int i = 0; i < {M}; i++) {{
+    for (int j = 0; j < {K} / 4; j++) {{
+      read_and_pad(&aa[i*A_stride + j*4], (int32_t*) &aa_pad[i*{K} + j*4], (int32_t*) &aa_pad[i*{K} + j*4 + 2]);
+    }}
+  }}
 
-  for (int i = 0; i < {N}; i++)
-    for (int j = 0; j < {K} / 4; j++)
+  for (int i = 0; i < {N}; i++) {{
+    for (int j = 0; j < {K} / 4; j++) {{
       read_and_pad(&bb[i*B_stride + j*4], (int32_t*) &bb_pad[i*{K} + j*4], (int32_t*) &bb_pad[i*{K} + j*4 + 2]);
-
-  for (int i = 0; i < {M}; i++) {{
-    int16_t aa_pad_line[{K}];
-    for (int l = 0; l < {K} / 4; l++)
-      read_and_pad(&aa[i*A_stride + l*4], (int32_t*) &aa_pad_line[l*4], (int32_t*) &aa_pad_line[l*4 + 2]);
-
-    for (int j = 0; j < {N}; j++) {{
-      int32_t *aa_ptr = (int32_t *) aa_pad_line;
-      int32_t *bb_ptr = (int32_t *) &bb_pad[j*{K}];
-      int32_t sum = 0;
-      for (int l = 0; l < 2 * ({K} / 4); l++) {{
-        sum = __SMLAD(*aa_ptr, *bb_ptr, sum);
-        ++ aa_ptr; ++ bb_ptr;
-      }}
-      cc[i*C_stride + j] += sum;
     }}
   }}
 
-  if ( {K} % 4 != 0 )
-    gemm_{M}x{N}_update_rest_{uniq_id}({K}, aa, bb, cc, A_stride, B_stride, C_stride);
-
-  return 0;
-}}
-
-
-
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm16_{M}x{N}_body_rest_{uniq_id}(
-    int K,
-    int16_t *aa, int16_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{
-  int k_base = (K / 2) * 2;
   for (int i = 0; i < {M}; i++) {{
     for (int j = 0; j < {N}; j++) {{
-      int16_t *a_ptr = &aa[i * A_stride + k_base];
-      int16_t *b_ptr = &bb[j * B_stride + k_base];
-      cc[i * C_stride + j] = (int32_t) a_ptr[0] * (int32_t) b_ptr[0];
-    }}
-  }}
-  return 0;
-}}
-
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm16_{M}x{K}x{N}_body_loop_{uniq_id}(
-    int16_t *aa, int16_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{
-  for (int i = 0; i < {M}; i++) {{
-    for (int j = 0; j < {N}; j++) {{
-      int32_t sum = 0;
-      for (int l = 0; l < {K}; l++) {{
-        sum += (int32_t) aa[i*A_stride + l] * (int32_t) bb[j*B_stride + l];
-      }}
-      // NOTE: this is the line where `*_body` differs from `*_update`. here
-      // we're *setting* the result, instead of accumulating, because we know
-      // the `i` and `j` itervars span their entire respective axes.
-      cc[i*C_stride + j] = sum;
-    }}
-  }}
-  return 0;
-}}
-
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm16_{M}x{K}x{N}_body_{uniq_id}(
-    int16_t *aa, int16_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{  
-  if ( {M} < 2 || {N} < 2 )
-    return gemm16_{M}x{K}x{N}_body_loop_{uniq_id}(aa, bb, cc, A_stride, B_stride, C_stride);  
-
-  for (int i = 0; i < {M}; i++) {{
-    for (int j = 0; j < {N}; j++) {{
-      int32_t *aa_ptr = (int32_t *) &aa[i*A_stride];
-      int32_t *bb_ptr = (int32_t *) &bb[j*B_stride];
-    
       int32_t sum = 0;
       for (int l = 0; l < {K} / 2; l++) {{
-        sum = __SMLAD(*aa_ptr, *bb_ptr, sum);
-        ++ aa_ptr; ++ bb_ptr;
-      }}
-      // NOTE: this is the line where `*_body` differs from `*_update`. here
-      // we're *setting* the result, instead of accumulating, because we know
-      // the `i` and `j` itervars span their entire respective axes.
-      cc[i*C_stride + j] = sum;
-    }}
-  }}
-
-  if ( {K} % 2 != 0 )
-    gemm16_{M}x{N}_body_rest_{uniq_id}({K}, aa, bb, cc, A_stride, B_stride, C_stride);
-
-  return 0;
-}}
-
-
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm16_{M}x{N}_update_rest_{uniq_id}(
-    int K,
-    int16_t *aa, int16_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{
-  int k_base = (K / 2) * 2;
-  for (int i = 0; i < {M}; i++) {{
-    for (int j = 0; j < {N}; j++) {{
-      int16_t *a_ptr = &aa[i * A_stride + k_base];
-      int16_t *b_ptr = &bb[j * B_stride + k_base];
-      cc[i * C_stride + j] += (int32_t) a_ptr[0] * (int32_t) b_ptr[0];
-    }}
-  }}
-  return 0;
-}}
-
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm16_{M}x{K}x{N}_update_loop_{uniq_id}(
-    int16_t *aa, int16_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{
-  for (int i = 0; i < {M}; i++) {{
-    for (int j = 0; j < {N}; j++) {{
-      int32_t sum = 0;
-      for (int l = 0; l < {K}; l++) {{
-        sum += (int32_t) aa[i*A_stride + l] * (int32_t) bb[j*B_stride + l];
-      }}
-      cc[i*C_stride + j] += sum;
-    }}
-  }}
-  return 0;
-}}
-
-#ifdef __cplusplus
-extern "C"
-#endif
-__STATIC_FORCEINLINE int32_t gemm16_{M}x{K}x{N}_update_{uniq_id}(
-    int16_t *aa, int16_t *bb, int32_t *cc,
-    int A_stride, int B_stride, int C_stride) {{  
-  if ( {M} < 2 || {N} < 2 )
-    return gemm16_{M}x{K}x{N}_update_loop_{uniq_id}(aa, bb, cc, A_stride, B_stride, C_stride);  
-
-  for (int i = 0; i < {M}; i++) {{
-    for (int j = 0; j < {N}; j++) {{
-      int32_t *aa_ptr = (int32_t *) &aa[i*A_stride];
-      int32_t *bb_ptr = (int32_t *) &bb[j*B_stride];
-
-      int32_t sum = 0;
-      for (int l = 0; l < {K} / 2; l++) {{
-        sum = __SMLAD(*aa_ptr, *bb_ptr, sum);
-        ++ aa_ptr; ++ bb_ptr;
+        sum = __SMLAD(
+          *((int32_t*) &aa_pad[i*{K} + l*2]),
+          *((int32_t*) &bb_pad[j*{K} + l*2]),
+          sum);
       }}
       cc[i*C_stride + j] += sum;
     }}
   }}
 
-  if ( {K} % 2 != 0 )
-    gemm16_{M}x{N}_update_rest_{uniq_id}({K}, aa, bb, cc, A_stride, B_stride, C_stride);
-
   return 0;
 }}
-
-
 
 #ifdef __cplusplus
 extern "C"
