+++ tests/micro/zephyr/test_zephyr_aot.py
@@ -15,9 +15,11 @@
 # specific language governing permissions and limitations
 # under the License.
 
+import io
 import logging
 import os
 import sys
+import logging
 import pathlib
 import tarfile
 import tempfile
@@ -30,17 +32,104 @@ import tvm.rpc
 import tvm.micro
 from tvm.micro.project_api import server
 import tvm.testing
-from tvm import relay
+import tvm.relay as relay
 
+from tvm.contrib import utils
 from tvm.contrib.download import download_testdata
 from tvm.micro.interface_api import generate_c_interface_header
 
 import conftest
-from test_utils import create_header_file
-from test_utils import build_project
-from test_utils import get_message
 
 _LOG = logging.getLogger(__name__)
+
+
+def _build_project(temp_dir, zephyr_board, west_cmd, mod, build_config, extra_files_tar=None):
+    template_project_dir = (
+        pathlib.Path(__file__).parent
+        / ".."
+        / ".."
+        / ".."
+        / "apps"
+        / "microtvm"
+        / "zephyr"
+        / "template_project"
+    ).resolve()
+    project_dir = temp_dir / "project"
+    project = tvm.micro.generate_project(
+        str(template_project_dir),
+        mod,
+        project_dir,
+        {
+            "extra_files_tar": extra_files_tar,
+            "project_type": "aot_demo",
+            "west_cmd": west_cmd,
+            "verbose": bool(build_config.get("debug")),
+            "zephyr_board": zephyr_board,
+        },
+    )
+    project.build()
+    return project, project_dir
+
+
+def _create_header_file(tensor_name, npy_data, output_path, tar_file):
+    """
+    This method generates a header file containing the data contained in the numpy array provided.
+    It is used to capture the tensor data (for both inputs and expected outputs).
+    """
+    header_file = io.StringIO()
+    header_file.write("#include <stddef.h>\n")
+    header_file.write("#include <stdint.h>\n")
+    header_file.write("#include <dlpack/dlpack.h>\n")
+    header_file.write(f"const size_t {tensor_name}_len = {npy_data.size};\n")
+
+    if npy_data.dtype == "int8":
+        header_file.write(f"int8_t {tensor_name}[] =")
+    elif npy_data.dtype == "int32":
+        header_file.write(f"int32_t {tensor_name}[] = ")
+    elif npy_data.dtype == "uint8":
+        header_file.write(f"uint8_t {tensor_name}[] = ")
+    elif npy_data.dtype == "float32":
+        header_file.write(f"float {tensor_name}[] = ")
+    else:
+        raise ValueError("Data type not expected.")
+
+    header_file.write("{")
+    for i in np.ndindex(npy_data.shape):
+        header_file.write(f"{npy_data[i]}, ")
+    header_file.write("};\n\n")
+
+    header_file_bytes = bytes(header_file.getvalue(), "utf-8")
+    raw_path = pathlib.Path(output_path) / f"{tensor_name}.h"
+    ti = tarfile.TarInfo(name=str(raw_path))
+    ti.size = len(header_file_bytes)
+    ti.mode = 0o644
+    ti.type = tarfile.REGTYPE
+    tar_file.addfile(ti, io.BytesIO(header_file_bytes))
+
+
+def _read_line(fd, timeout_sec: int):
+    data = ""
+    new_line = False
+    while True:
+        if new_line:
+            break
+        new_data = fd.read(1, timeout_sec=timeout_sec)
+        logging.debug(f"read data: {new_data}")
+        for item in new_data:
+            new_c = chr(item)
+            data = data + new_c
+            if new_c == "\n":
+                new_line = True
+                break
+    return data
+
+
+def _get_message(fd, expr: str, timeout_sec: int):
+    while True:
+        data = _read_line(fd, timeout_sec)
+        logging.debug(f"new line: {data}")
+        if expr in data:
+            return data
 
 
 @tvm.testing.requires_micro
@@ -103,12 +192,12 @@ def test_tflite(temp_dir, board, west_cmd, tvm_debug):
                 )
                 tf.add(header_path, arcname=os.path.relpath(header_path, tar_temp_dir))
 
-            create_header_file("input_data", sample, "include", tf)
-            create_header_file(
+            _create_header_file("input_data", sample, "include", tf)
+            _create_header_file(
                 "output_data", np.zeros(shape=output_shape, dtype="float32"), "include", tf
             )
 
-        project, _ = build_project(
+        project, _ = _build_project(
             temp_dir,
             board,
             west_cmd,
@@ -120,9 +209,9 @@ def test_tflite(temp_dir, board, west_cmd, tvm_debug):
     project.flash()
     with project.transport() as transport:
         timeout_read = 60
-        get_message(transport, "#wakeup", timeout_sec=timeout_read)
+        _get_message(transport, "#wakeup", timeout_sec=timeout_read)
         transport.write(b"start\n", timeout_sec=5)
-        result_line = get_message(transport, "#result", timeout_sec=timeout_read)
+        result_line = _get_message(transport, "#result", timeout_sec=timeout_read)
 
     result_line = result_line.strip("\n")
     result_line = result_line.split(":")
@@ -164,10 +253,10 @@ def test_qemu_make_fail(temp_dir, board, west_cmd, tvm_debug):
                     lowered.libmod_name, ["input_1"], ["output"], model_files_path
                 )
                 tf.add(header_path, arcname=os.path.relpath(header_path, tar_temp_dir))
-            create_header_file("input_data", np.zeros(shape=shape, dtype=dtype), "include", tf)
-            create_header_file("output_data", np.zeros(shape=shape, dtype=dtype), "include", tf)
+            _create_header_file("input_data", np.zeros(shape=shape, dtype=dtype), "include", tf)
+            _create_header_file("output_data", np.zeros(shape=shape, dtype=dtype), "include", tf)
 
-        project, project_dir = build_project(
+        project, project_dir = _build_project(
             temp_dir,
             board,
             west_cmd,
